fix: remove additional dependency cycles
