+++ .changeset/grumpy-numbers-bake.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react": patch
+---
+
+Fix internal dependency cycles
