+++ packages/react/src/api/ThreadRuntime.ts
@@ -27,7 +27,7 @@ import {
   ThreadListItemRuntimePath,
   ThreadRuntimePath,
 } from "./RuntimePathTypes";
-import { ThreadListItemState } from "./ThreadListItemRuntime";
+import type { ThreadListItemState } from "./RuntimeBindings";
 import { RunConfig } from "../types/AssistantTypes";
 import { EventSubscriptionSubject } from "./subscribable/EventSubscriptionSubject";
 import { symbolInnerMessage } from "../runtimes/external-store/getExternalStoreMessage";
