+++ packages/react/src/types/AssistantTypes.ts
@@ -1,75 +1,32 @@
-import { CompleteAttachment } from "./AttachmentTypes";
-import { ReadonlyJSONObject, ReadonlyJSONValue } from "assistant-stream/utils";
+import type { CompleteAttachment } from "./AttachmentTypes";
+import { ReadonlyJSONValue } from "assistant-stream/utils";
+import type {
+  TextMessagePart,
+  ReasoningMessagePart,
+  SourceMessagePart,
+  ImageMessagePart,
+  FileMessagePart,
+  Unstable_AudioMessagePart,
+  ToolCallMessagePart,
+  ThreadUserMessagePart,
+  ThreadAssistantMessagePart,
+} from "./MessagePartTypes";
 
+// Re-export message part types for convenience
+export type {
+  TextMessagePart,
+  ReasoningMessagePart,
+  SourceMessagePart,
+  ImageMessagePart,
+  FileMessagePart,
+  Unstable_AudioMessagePart,
+  ToolCallMessagePart,
+  ThreadUserMessagePart,
+  ThreadAssistantMessagePart,
+};
+
+// Alias for the role of a thread message
 export type MessageRole = ThreadMessage["role"];
-
-export type TextMessagePart = {
-  readonly type: "text";
-  readonly text: string;
-  readonly parentId?: string;
-};
-
-export type ReasoningMessagePart = {
-  readonly type: "reasoning";
-  readonly text: string;
-  readonly parentId?: string;
-};
-
-export type SourceMessagePart = {
-  readonly type: "source";
-  readonly sourceType: "url";
-  readonly id: string;
-  readonly url: string;
-  readonly title?: string;
-  readonly parentId?: string;
-};
-
-export type ImageMessagePart = {
-  readonly type: "image";
-  readonly image: string;
-};
-
-export type FileMessagePart = {
-  readonly type: "file";
-  readonly data: string;
-  readonly mimeType: string;
-};
-
-export type Unstable_AudioMessagePart = {
-  readonly type: "audio";
-  readonly audio: {
-    readonly data: string;
-    readonly format: "mp3" | "wav";
-  };
-};
-
-export type ToolCallMessagePart<
-  TArgs = ReadonlyJSONObject,
-  TResult = unknown,
-> = {
-  readonly type: "tool-call";
-  readonly toolCallId: string;
-  readonly toolName: string;
-  readonly args: TArgs;
-  readonly result?: TResult | undefined;
-  readonly isError?: boolean | undefined;
-  readonly argsText: string;
-  readonly artifact?: unknown;
-  readonly parentId?: string;
-};
-
-export type ThreadUserMessagePart =
-  | TextMessagePart
-  | ImageMessagePart
-  | FileMessagePart
-  | Unstable_AudioMessagePart;
-
-export type ThreadAssistantMessagePart =
-  | TextMessagePart
-  | ReasoningMessagePart
-  | ToolCallMessagePart
-  | SourceMessagePart
-  | FileMessagePart;
 
 type MessageCommonProps = {
   readonly id: string;
